requirements.txt
customtkinter==5.2.1
Pillow==10.2.0
openpyxl==3.1.2
pandas==2.2.2
PyPDF2==3.0.1
pdfplumber>=0.7.6
reportlab==4.1.0
pyinstaller==5.13.0